Freetype integrated to assets convertor
