+++ GameEngine/GameEngine.sln
@@ -110,7 +110,6 @@ Global
 		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
 		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.Debug|Win32.ActiveCfg = Debug|Win32
-		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.Debug|Win32.Build.0 = Debug|Win32
 		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.MinSizeRel|Any CPU.ActiveCfg = Release|Win32
 		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.MinSizeRel|Mixed Platforms.ActiveCfg = Release|Win32
 		{60ACE0BC-C15B-4A41-A735-63FD967CC91F}.MinSizeRel|Mixed Platforms.Build.0 = Release|Win32
@@ -130,7 +129,6 @@ Global
 		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
 		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.Debug|Mixed Platforms.Build.0 = Debug|Win32
 		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.Debug|Win32.ActiveCfg = Debug|Win32
-		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.Debug|Win32.Build.0 = Debug|Win32
 		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.MinSizeRel|Any CPU.ActiveCfg = Release|Win32
 		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.MinSizeRel|Mixed Platforms.ActiveCfg = Release|Win32
 		{CE40F8DC-055E-424B-A0D0-1C020EB3632F}.MinSizeRel|Mixed Platforms.Build.0 = Release|Win32
@@ -169,7 +167,6 @@ Global
 		{3073A470-6943-400E-BDFF-D25AA01577F8}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
 		{3073A470-6943-400E-BDFF-D25AA01577F8}.Debug|Mixed Platforms.Build.0 = Debug|Win32
 		{3073A470-6943-400E-BDFF-D25AA01577F8}.Debug|Win32.ActiveCfg = Debug|Win32
-		{3073A470-6943-400E-BDFF-D25AA01577F8}.Debug|Win32.Build.0 = Debug|Win32
 		{3073A470-6943-400E-BDFF-D25AA01577F8}.MinSizeRel|Any CPU.ActiveCfg = Release|Win32
 		{3073A470-6943-400E-BDFF-D25AA01577F8}.MinSizeRel|Mixed Platforms.ActiveCfg = Release|Win32
 		{3073A470-6943-400E-BDFF-D25AA01577F8}.MinSizeRel|Mixed Platforms.Build.0 = Release|Win32
@@ -188,7 +185,6 @@ Global
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.Debug|Win32.ActiveCfg = Debug|Win32
-		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.Debug|Win32.Build.0 = Debug|Win32
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.MinSizeRel|Any CPU.ActiveCfg = Release|Win32
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.MinSizeRel|Mixed Platforms.ActiveCfg = Release|Win32
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.MinSizeRel|Mixed Platforms.Build.0 = Release|Win32
@@ -206,6 +202,7 @@ Global
 		{ECFC3CA5-DB31-4109-8A00-88358E23B3B9}.RelWithDebInfo|Win32.Build.0 = Release|Win32
 		{EB423B39-F564-498E-B291-0D9676AF09B5}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{EB423B39-F564-498E-B291-0D9676AF09B5}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
+		{EB423B39-F564-498E-B291-0D9676AF09B5}.Debug|Mixed Platforms.Build.0 = Debug|Win32
 		{EB423B39-F564-498E-B291-0D9676AF09B5}.Debug|Win32.ActiveCfg = Debug|Win32
 		{EB423B39-F564-498E-B291-0D9676AF09B5}.Debug|Win32.Build.0 = Debug|Win32
 		{EB423B39-F564-498E-B291-0D9676AF09B5}.MinSizeRel|Any CPU.ActiveCfg = Release|Win32
@@ -225,9 +222,7 @@ Global
 		{EB423B39-F564-498E-B291-0D9676AF09B5}.RelWithDebInfo|Win32.Build.0 = Release|Win32
 		{15D1542F-C88C-4549-8029-9D4495E54A04}.Debug|Any CPU.ActiveCfg = Debug|Win32
 		{15D1542F-C88C-4549-8029-9D4495E54A04}.Debug|Mixed Platforms.ActiveCfg = Debug|Win32
-		{15D1542F-C88C-4549-8029-9D4495E54A04}.Debug|Mixed Platforms.Build.0 = Debug|Win32
 		{15D1542F-C88C-4549-8029-9D4495E54A04}.Debug|Win32.ActiveCfg = Debug|Win32
-		{15D1542F-C88C-4549-8029-9D4495E54A04}.Debug|Win32.Build.0 = Debug|Win32
 		{15D1542F-C88C-4549-8029-9D4495E54A04}.MinSizeRel|Any CPU.ActiveCfg = Release|Win32
 		{15D1542F-C88C-4549-8029-9D4495E54A04}.MinSizeRel|Mixed Platforms.ActiveCfg = Release|Win32
 		{15D1542F-C88C-4549-8029-9D4495E54A04}.MinSizeRel|Mixed Platforms.Build.0 = Release|Win32
