+++ GameEngine/GameEngine/Engine/Text/FontGlyph.hh
@@ -12,8 +12,6 @@
 
 #include <array>
 #include <cstddef>
-#include <ft2build.h>
-#include FT_FREETYPE_H
 
 struct FontGlyph
 {
